Add some more tests for the cluster manager.
